[🧹chore] lib/supabase/middleware.ts 파일 재정의 및 파일 위치 이동
- app/middleware.ts 상위로 파일 이동
